organized app.vue so variable definitions and auth functions are at the top, other functions at the bottom changed isOpen to isNavOpen and added isPreferencesOpen for user preferences screen
Updated logged in block to have all the init logic

set username as well, only use the first name in the welcome message

Added all post-load/async calls to the client only block

new postUUID function to get the UUID from the server, and a callback to kick off any logic that requires the UUID.

Removed getNuxtData, as it's session only, and started using localStorage everywhere instead.

new theme logic- functions are: setTheme (in DB), getTheme (from DB), and setClientTheme (sets the theme on the client)

Updated button variants from none to ghost as none isn't correct

Added slideover for user preferences menu, with a few buttons for logging in/out and themes

Themes are only "glassy" and "default" for now. Default needs to reload the page when it's set.

Removed the glassy styles from app and put them in a stylesheet in the public dir so we can access them during runtime.

Updated clicker and envelopes pages to use localStorage instead of nuxt data as well.

Fixed the three line bars in home page text.

Added getTheme and setTheme API end points.

Updated checkValidKey util to also check if the key is "guest". If so, this fails, so no requests to/from the server will work in guest mode.