Removed unused import
new selectedItemsAccordion ref that only serves to represent the accordion that hides the selected items table

Removed zombie code

Re-arranged table columns, new column "selected"

new item quantity defaults to 0

added selected field on items

removed all debug messages

updated the select function so it can toggle the selected field on items when the item is selected. This also adds it to a second array, selectedItems, and is displayed in a second list.

The select function also has a delay of 2 seconds before removing an item from the list.

new function filterSlectedItems which will update a new array filteredItems. This array is bound to our shopping list table and only items in it that have their selected flags as false will be returned.

Formatted the document so lots of HTML changes but no many real changes.

Added accordion, updated the table, added second table.