Added drag and drop lib
imported uuid and drag and drop to grocery list.

initialized an empty drag and drop list.

Added list of departments.

Every time we fire showEdit, set the list of draggable items to the current array of items.

All item id's are now uuids.

Updated deleteItem so it removes the item from the selected list as well as the edit list.

Added deleteSelectedItems, which clears the selectedItem array, and deletes selected items from the edit items array.

Updated select function to use IDs isntead of the index.

Added a demo for the sortable list above the edit item list. Will need to update the edit item components so they are <li> in a <ul>.

Added a button that will be used to drag and drop edit items onto the edit item component.

Added the linethrough class to the edit item input fields so they are striked out when selected.

Updated pricing display in shopping mode. Multiplies by quantity, unless quantity is less than 1.

Added strikethrough class to department column.

Added the button to clear out the selected items list.

Wrapped edit items in formGroups.

Replaced the item name input with a textarea.

Replaced department field with a dropdown list.

Put the submit item button in the footer.