Imported slicksort, create app, and SlickList and SlickItem.
Removed simple arrays being used for testing.

New function reOrderListItems that accepts the event from the sort-end event. It then removes the item at the old index, and re-inserts it into the list at the newIndex.

Added a slicklist above the item edit list. This is to keep the list item's simple while figuring out how it works exactly.

This works, I just need to now wrap that slick list around the item edit buttons and fields themselves, and then set the drag handle to the chevron arrows.