Removed unused import for the slicksort plugin. We don't need this.
Imported DragHandle component as well.

Moved log function to a util so I can use it wherever.

Removed the ugly reOrderListItem function, as it turns out I didn't need to do any of that. Instead, when the @update:list event fires from Slicksort and returns the NEWLY SORTED ARRAY, I just set it in sortTheList and call set and filter.

Updated the Slicklist a bit and left a comment for later about why I can't see the items being moved (it's cause we're in an overlay)