added hidden class to all table headers except name
add some more validation for filteredSelectedItems so we don't hit undefined item name errors on empty lists

manually find the name column and set it's attribute to colspan 4 so just the name header is the full width of the column. This will be our 'utils' scrolling container.

Updated the table's base styles.

Added a header template for the name column. Inside we for each through all the columns and build buttons for them. This is where all table related actions will go, i.e. sorting, sharing.

Updated the leave shopping mode button so it's the full screen width when not in focus mode.