app.vue updates: import getUserPrefsJSON function
new function isUUIDSet, quick util function to check if we have a UUID in the cache or not.

in postUUIDCallback, call setUserPrefs.

New function setUserPrefs, will check cache for userPrefs, otherwise queries the DB for them and sets them in cache from the response from the server

Updated setClientTheme to grab data from userPrefs, updated logic a bit to clean it up.

Removed getTheme function.

Updated setTheme function to grab userPrefs, updated logic to clean it up a bit.

updated onThemeSelect to grab from userPrefs instead.

Updated 'on logged in and on client' init block to check isUUIDSet, then run setUserPrefs, or postUUID if UUID is not set, which calls setUserPrefs and setClientTheme.

Updated 'not logged in' init block to call setUserPrefs, still works when only in 'cache only' mode.

package updates

updated nuxt pwa to 0.6.0 hope nothing breaks
added supabase-js

added page for /.more as a placeholder

UUID.ts updates:
import supabase.
Check to see if user is in db. If they are, send back that UUID.
If they are not, insert the email/uuid to the db.

new endpoint getPrefs and setPrefs. Handles getting and setting user preferences.

new util function geteUserPrefsJSON, just grabs the user prefs object from the cache, validates it, and parses it and return sit as a JSON object.

new util function supabase.ts that creates a supabase client instance and returns it. Doesn't access the config directly, requires one to be passed.