Updated setUserPrefs to be a bit simpler. If no UUID and nothing in the cache, we set a default value for the user prefs.
Updated setBudget in clicker. Now, simply set's the value of the current budget to the user prefs obj in cache.

Updated getBudget in clicker (was getUserBudget). Now, simply gets the budget from user prefs obj in cache, if not set in cache yet defaults to 100.

Updated getRecords in clicker to return if no UUID is set.

Updated initialize in clicker to run getRecords and getBudget, even if we're not logged in. Moving towards cache-first build.

Updated getBudget in shopping list to match clicker's new getBudget.

Removed get and set Budget endpoints, no longer stored as individual values in a database.

Updated getPrefs endpoint to use checkValidKey and request clickerBudget as well.

Updated checkValidKey to check key length instead of it it is 'guest', it will never be guest.

Updated getUserPrefsJSON util function to give the clickerBudget an initial value of 100 if we use the default.