Reorganized app.vue's layout
Updated setUserPrefs to be getUserPrefs, it was very missleading before. Cleaned up the logic. We're taking a DB first approach.

Installed VueUse to utilize useOnline to check online status.

Removed calls to other functions from within functions that caused spaghetti code in app.vue

updated setClient theme to literally just set the client theme, nothing more.

upder setTheme to setUserPrefsThemeField, a bit more descriptive. It updates the db if we can, along with the cache, and calls setClientTheme (only time we're allowed to call another internal function, as this is an 'action' function triggered from a button)

New function setClientUserData, which just takes the if block from init and places it here to clean up the init logic.

Cleaned up the main init logic to be much more clear and less redundant.

Added vueuse to nuxt config.

Added a todo to setBudget

Add a new util function checkUUID. Can be used wherever we were previoulsy checking for a valid UUID.