New util function getNewestTimestamp, just returns the name of the source that's newer (cache or db)
Import axios to use for requests

New types folder and type UserPrefs.

Overhauled getUserPrefs again to first grab data from the cache, if it's not empty convert to JSON, if we can get from DB, check which is newer, if DB is newer we set the cache to DB, if cache is newer we set DB to cache, updating the timestamp when appropriate, if response is invalid we default to whatever was in the cache or use the defaults if cache is empty, and finally if we can't check DB, just user whatever is in the cache.

New function setUserPrefs to update the user prefs rec in the DB with the object that was passed to it, and update the timestamp on the object.

Removed calls to setPrefs and replaced with the function setUserPrefs.

Commented out clearUserData till we have a working one.

Updated getUserPrefs JSON to use the UserPrefs type. Also included the createdAt timestamp for defaults.

New util function update timestamp, I probably won't use it but keeping it because it's nice.